introduce fill buffer generator
