+++ bazel/core.bzl
@@ -613,6 +613,7 @@ MLN_CORE_SOURCE = [
     "src/mbgl/util/quaternion.cpp",
     "src/mbgl/util/quaternion.hpp",
     "src/mbgl/gfx/polyline_generator.cpp",
+    "src/mbgl/gfx/fill_generator.cpp",
     "src/mbgl/util/rapidjson.cpp",
     "src/mbgl/util/rapidjson.hpp",
     "src/mbgl/util/rect.hpp",
@@ -820,6 +821,7 @@ MLN_CORE_HEADERS = [
     "include/mbgl/util/noncopyable.hpp",
     "include/mbgl/util/platform.hpp",
     "include/mbgl/gfx/polyline_generator.hpp",
+    "include/mbgl/gfx/fill_generator.hpp",
     "include/mbgl/util/premultiply.hpp",
     "include/mbgl/util/projection.hpp",
     "include/mbgl/util/range.hpp",
